backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# Background Tasks
celery==5.4.0
redis==5.2.1

# AI Integration
openai==1.58.1
anthropic==0.42.0

# Log Parsing & Analysis
pandas==2.2.3
python-dateutil==2.9.0.post0
regex==2024.11.6
ipaddress==1.0.23
user-agents==2.2.0
geoip2==4.8.1

# Search (Elasticsearch)
elasticsearch==8.17.0

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
aiofiles==24.1.0
jinja2==3.1.5
orjson==3.10.13

# Export & Reporting
reportlab==4.2.5
python-docx==1.1.2

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
ruff==0.8.6
